+++ src/LaravelServiceProvider.php
@@ -26,6 +26,7 @@ use Upmind\ProvisionProviders\DomainNames\Ricta\Provider as Ricta;
 use Upmind\ProvisionProviders\DomainNames\UGRegistry\Provider as UGRegistry;
 use Upmind\ProvisionProviders\DomainNames\DomainNameApi\Provider as DomainNameApi;
 use Upmind\ProvisionProviders\DomainNames\CentralNic\Provider as CentralNic;
+use Upmind\ProvisionProviders\DomainNames\GoDaddy\Provider as GoDaddy;
 use Upmind\ProvisionProviders\DomainNames\CentralNicReseller\Provider as CentralNicReseller;
 
 class LaravelServiceProvider extends ProvisionServiceProvider
@@ -55,6 +56,7 @@ class LaravelServiceProvider extends ProvisionServiceProvider
         $this->bindProvider('domain-names', 'domain-name-api', DomainNameApi::class);
         $this->bindProvider('domain-names', 'namecheap', Namecheap::class);
         $this->bindProvider('domain-names', 'centralnic', CentralNic::class);
+        $this->bindProvider('domain-names', 'godaddy', GoDaddy::class);
         $this->bindProvider('domain-names', 'centralnic-reseller', CentralNicReseller::class);
     }
 }